Moved SyncCommandParser to correct dir
